Simplifying stuff, but srt doesn't work right now
