+++ examples/evp.py
@@ -1,4 +1,5 @@
 from seaice import *
+from firedrake import *
 
 # TEST 2 : EVP
 
@@ -8,12 +9,18 @@ timescale = 10
 
 dirname = "./output/EVP/u_timescale={}_timestep={}.pvd".format(timescale, timestep)
 
+number_of_triangles = 35
+length = 5 * 10 ** 5
+mesh = SquareMesh(number_of_triangles, number_of_triangles, length)
+rheology = 'VP'
+
 timestepping = TimesteppingParameters(timescale=timescale, timestep=timestep)
 output = OutputParameters(dirname=dirname, dumpfreq=dumpfreq)
 solver = SolverParameters()
 params = SeaIceParameters()
 
-evp = Evp(number_of_triangles=35, params=params, timestepping=timestepping, output=output, solver_params=solver)
+evp = SeaIceModel(mesh=mesh, length=length, rheology=rheology, timestepping=timestepping, output=output, params=params,
+                  solver_params=solver)
 
 t = 0
 while t < timescale - 0.5 * timestep:
